ci: replace populate/dump with ingest on docker data
