+++ packages/code-du-travail-data/README.md
@@ -6,48 +6,18 @@ L'image docker DATA contient un dump des documents ainsi que les vecteurs associ
 
 L'ensemble des documents et données utilisés par l'api est stocké dans une base elasticsearch et peut être mis à jour en lançant le script d'indexation sans avoir à re-déployer l'api ou le frontend. Cela à l'exception de la liste des courriers qui fait toujours parti du dataset et des mots du glossaire qui sont importés dans le frontend pour permettre d'afficher des tooltips dynamiques.
 
-Le processus de mise à jour des données est composé de 2 etapes.
+### Indexation des document dans elasticsearch
 
-### Exportation
+Les données utilisées par le site sont issues de packages de données (@socialgouv/\*-data).
 
-Les données utilisées par le site sont issues de packages de données (@socialgouv/*-data). Ces données sont compilées dans un fichier de dump dans un format pret pour l'indexation. Pour créer ce fichier, on utilise la commande.
-
-```sh
-$ yarn workspace @cdt/data dump-dev
-```
-
-Pour chaque document, il possible aussi de rajouter une représentation vectorielle qui servira à la recherche semantique. Pour cela, il faut que la variable d'environnement NLP_URL soit fournie et pointe vers une instance du service NLP. Reportez-vous au [README.md du projet serving-ml](https://github.com/SocialGouv/serving-ml) pour voir comment démarrer une instance locale du service. Une instance est également accessible à cette URL : https://preprod-serving-ml.dev2.fabrique.social.gouv.fr/ 
+Pour chaque document, il possible aussi de rajouter une représentation vectorielle qui servira à la recherche semantique. Pour cela, il faut que la variable d'environnement NLP_URL soit fournie et pointe vers une instance du service NLP. Reportez-vous au [README.md du projet serving-ml](https://github.com/SocialGouv/serving-ml) pour voir comment démarrer une instance locale du service. Une instance est également accessible à cette URL : https://preprod-serving-ml.dev2.fabrique.social.gouv.fr/
 
 ```sh
 # Exporter les données vers elacticsearch avec leur representation vectorielle (title_vector)
-$ NLP_URL=https://preprod-serving-ml.dev2.fabrique.social.gouv.fr yarn workspace @cdt/data dump-dev
+$ NLP_URL=https://preprod-serving-ml.dev2.fabrique.social.gouv.fr yarn workspace @cdt/data ingest
 ```
 
-### Importation dans elasticsearh
-
-Une fois le fichier de dump généré, on utilise la commande 
-
-```sh
-$ yarn workspace @cdt/data populate-dev
-```
-pour créer 2 index elasticsearch( un dédié aux documents, l'autre aux suggestions).
-
-
-Par defaut, le script d'import va chercher les données dans `/packages/code-du-travail-data/dist/dump.data.json` mais il possible de changer ce chemin avec la variable d'environnement `DUMP_PATH`.
-
-
-### Récupération depuis l'image master
-
-Si vous disposez d'une image nlp sur votre machine vous pouvez copier le fichier de dump complet
-via la commande docker suivante.
-
-```sh
-docker run \
-   --rm --entrypoint cat \
-   registry.gitlab.factory.social.gouv.fr/socialgouv/code-du-travail-numerique/data:$(git rev-parse origin/master) \
-   /app/dist/dump.data.json \
-   >! packages/code-du-travail-data/dist/dump.tf.json
-```
+Le script va créer 2 index elasticsearch( un dédié aux documents, l'autre aux suggestions).
 
 Les données provenant de packages externes sont désormais mis à jour de manière automatique via le bot renovate qui maintient les versions de nos dépendances à jour, seuls les outils et les modèles nécessitent une mise à jour des deploiement de l'api et du frontend.
