+++ packages/code-du-travail-data/indexing/index.js
@@ -1,6 +1,10 @@
 import { Client } from "@elastic/elasticsearch";
 import { logger } from "@socialgouv/cdtn-logger";
+import { SOURCES } from "@socialgouv/cdtn-sources";
+import PQueue from "p-queue";
+import retry from "p-retry";
 
+import { cdtnDocumentsGen } from "./cdtnDocuments";
 import { documentMapping } from "./document.mapping";
 import {
   createIndex,
@@ -9,7 +13,9 @@ import {
   version,
 } from "./esClientUtils";
 import { DOCUMENTS, SUGGESTIONS } from "./esIndexName";
+import { fetchCovisits } from "./monolog";
 import { populateSuggestions } from "./suggestion";
+import { vectorizeDocument } from "./vectorizer";
 
 const ES_INDEX_PREFIX = process.env.ES_INDEX_PREFIX || "cdtn";
 
@@ -19,27 +25,60 @@ const SUGGEST_INDEX_NAME = `${ES_INDEX_PREFIX}_${SUGGESTIONS}`;
 const ELASTICSEARCH_URL =
   process.env.ELASTICSEARCH_URL || "http://localhost:9200";
 
-const DUMP_PATH = process.env.DUMP_PATH || "../dist/dump.data.json";
-
-logger.info(`ElasticSearch at ${ELASTICSEARCH_URL}`);
+const NLP_URL = process.env.NLP_URL;
 
 const esClientConfig = {
+  auth: { apiKey: process.env.ELASTICSEARCH_DATA_TOKEN },
   node: `${ELASTICSEARCH_URL}`,
 };
 
-switch (process.env.NODE_ENV) {
-  case "production":
-    esClientConfig.auth = {
-      password: process.env.ELASTICSEARCH_PWD,
-      username: process.env.ELASTICSEARCH_USER || "elastic",
-    };
-    break;
+const client = new Client(esClientConfig);
+
+export async function addVector(data) {
+  if (NLP_URL) {
+    if (!data.title) {
+      logger.error(`No title for document ${data.source} / ${data.slug}`);
+    }
+    const title = data.title || "sans titre";
+    await vectorizeDocument(title, data.text)
+      .then((title_vector) => {
+        if (title_vector.message) {
+          throw new Error(`error fetching ${data.title}`);
+        }
+        data.title_vector = title_vector;
+      })
+      .catch((err) => {
+        logger.error(`error fetching ${data.title}`, err.message);
+      });
+  }
+
+  return Promise.resolve(data);
 }
 
-const client = new Client(esClientConfig);
+// these sources do not need NLP vectorization
+const excludeSources = [
+  SOURCES.CDT,
+  SOURCES.GLOSSARY,
+  SOURCES.PREQUALIFIED,
+  SOURCES.HIGHLIGHTS,
+  SOURCES.SHEET_MT_PAGE,
+  SOURCES.VERSIONS,
+];
 
 async function main() {
   const ts = Date.now();
+  const nlpQueue = new PQueue({ concurrency: 5 });
+
+  const monologQueue = new PQueue({ concurrency: 20 });
+
+  logger.info(`using cdtn elasticsearch ${process.env.ELASTICSEARCH_URL}`);
+
+  if (NLP_URL) {
+    logger.info(`Using NLP service to retrieve tf vectors on ${NLP_URL}`);
+  } else {
+    logger.info(`NLP_URL not defined, semantic search will be disabled.`);
+  }
+
   await version({ client });
 
   // Indexing documents/search data
@@ -48,12 +87,33 @@ async function main() {
     indexName: `${DOCUMENT_INDEX_NAME}-${ts}`,
     mappings: documentMapping,
   });
-  const documents = require(DUMP_PATH);
-  await indexDocumentsBatched({
-    client,
-    documents,
-    indexName: `${DOCUMENT_INDEX_NAME}-${ts}`,
-  });
+
+  const t0 = Date.now();
+  for await (const { source, documents } of cdtnDocumentsGen()) {
+    logger.info(`› ${source}... ${documents.length} items`);
+
+    // add covisits using pQueue (there is a plan to change this : see #2915)
+    const pDocs = documents.map((doc) =>
+      monologQueue.add(() => fetchCovisits(doc))
+    );
+    let covisitDocuments = await Promise.all(pDocs);
+    await monologQueue.onIdle();
+    // add NLP vectors
+    if (!excludeSources.includes(source)) {
+      const pDocs = covisitDocuments.map((doc) =>
+        nlpQueue.add(() => retry(() => addVector(doc), { retries: 3 }))
+      );
+      covisitDocuments = await Promise.all(pDocs);
+      await nlpQueue.onIdle();
+    }
+    await indexDocumentsBatched({
+      client,
+      documents: covisitDocuments,
+      indexName: `${DOCUMENT_INDEX_NAME}-${ts}`,
+    });
+  }
+
+  logger.info(`done in ${(Date.now() - t0) / 1000} s`);
 
   // Indexing Suggestions
   await populateSuggestions(client, `${SUGGEST_INDEX_NAME}-${ts}`);
@@ -97,12 +157,14 @@ async function main() {
 }
 
 main().catch((response) => {
+  console.error(response);
   if (response.body) {
-    logger.error(response.meta.statusCode);
-    logger.error(response.name);
-    logger.error(response.meta.meta.request);
+    logger.error({ statusCode: response.meta.statusCode });
+    logger.error({ name: response.name });
+    logger.error({ request: response.meta.meta.request });
+    logger.error({ body: response.body });
   } else {
-    logger.error(response);
+    logger.error({ response });
   }
-  process.exit(-1);
+  process.exit(1);
 });
