+++ packages/code-du-travail-data/package.json
@@ -5,16 +5,13 @@
   "private": true,
   "main": "index.js",
   "scripts": {
-    "build": "yarn build:populate",
-    "build:populate": "ncc build ./indexing/index.js -o dist/populate",
+    "build": "ncc build ./indexing/index.js -o dist",
     "check-slugs": "DISABLE_GLOSSARY=true  node -r esm indexing/slugChecker",
     "download-suggestion": "curl -L $SUGGEST_DATA_URL -o ./dataset/suggestions.txt",
-    "dump-dev": "node -r esm dump.js > ./dist/dump.data.json",
     "lint": "eslint ./indexing/*",
-    "populate-dev": "node -r esm indexing/index.js",
-    "populate": "NODE_OPTIONS=--max_old_space_size=1536 NODE_ENV=production node ./dist/populate/index.js",
+    "start": "NODE_OPTIONS=--max_old_space_size=1536 NODE_ENV=production node ./dist/index.js",
     "populate-suggest": "node -r esm indexing/suggestion",
-    "ingest": "node -r esm indexing/ingest",
+    "ingest": "node -r esm indexing/index",
     "prepush": "lint-staged",
     "test": "BUFFER_SIZE=10 SUGGEST_FILE=./indexing/__tests__/suggestion_data_test.txt jest"
   },
